VkBotsLongPoll move +some fixes
